Implement freeze overlap support
Freeze overlap lets you limit the lower bound of a partition based on
the upper bound of the partition immediately preceeding it. It does this
by moving data that would violate the new boundary condition from the
partition in question to the limbo table.

If you have highly regular data with a very few outliers (for example,
you have a timestamp column that generally tracks a few minutes behind
current time, but very rarely has entries from a month ago), and you
want to take advantage of CBE when querying on this data, then freeze
overlap is what you want.